[fix refactor] graph abstraction and reorganization
